refactor: ammend modal component so it can now display non fullscreen modals
Refactored the modal component so it can be used to display a non fullscreen modal over the map. Map modal is now displayed to the left of the map in a similar way to Google Maps.
